refactor: rename providers/accuweather to providers/weather
