+++ merino/backends/accuweather.py
@@ -6,7 +6,7 @@ from pydantic import BaseModel
 
 from merino.backends.exceptions import BackendError
 from merino.middleware.geolocation import Location
-from merino.providers.accuweather import (
+from merino.providers.weather import (
     CurrentConditions,
     Forecast,
     Temperature,
