+++ merino/providers/weather.py
